requirements.txt
asgiref==3.5.0
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==0.5.0
Django==3.0.8
django-allauth==0.49.0
django-cors-headers==3.11.0
django-filter==21.1
django-model-utils==4.2.0
django-rest-auth==0.9.5
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
filelock==3.4.0
gunicorn==20.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.1.0
oauthlib==3.2.0
platformdirs==2.4.0
psycopg2==2.9.3
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
unicodecsv==0.13.0
uritemplate==4.1.1
urllib3==1.26.8
virtualenv==20.10.0
whitenoise==6.0.0
